Update SimpleLibrary.h
this variable need to be Unsigned int like millis(). Without this fix debounce timer not works properly

+++ 00-LibrariesUsed/SimpleLibrary/SimpleLibrary.h
@@ -99,7 +99,7 @@ class Button{
   int _debounceDelay;
   int _reading;
   int _pin;
-  int _lastDebounceTime = 0;
+  unsigned long _lastDebounceTime = 0;
   
   public:
   void Format(int p, int dd);
